Port to NeoForge 1.21.1 and Epic Fight 21.17.3.1

+++ src/main/java/net/okitsu/ysmepicfightcompat/ysmref/ModelPreviewRendererAlias.java
@@ -11,7 +11,7 @@ final class ModelPreviewRendererAlias {
     }
 
     public static void renderPlayerOverlay(GuiGraphics graphics, LocalPlayer player,
-                                           double x, double y, float scale, float yawOffset,
+                                           float x, float y, float scale, float yawOffset,
                                            int depth, float partialTick) {
         throw new UnsupportedOperationException("YSM Mapping API source alias");
     }
